Login functionality
Changed and added js code for logging in,

changed logout link to a button to work with js,

can log in but have problems with navigation:
for some reason on successful login when going to home page, the page reloads itself constantly and there are no loops anywhere in the code. will keep work on it.