More progress
Assistants can now add patients,
patients list diaplays actual patients from DB,
When creating new users by admin or assistant a new firebase account is created as well as database user allowing the created user to sign in with email and password,
added another page for admin - admin/selected -> need to figure out how to lock on to a specific user using the firebase uid to be able to display eddit and delete the said user.
tables for admin and assistant are now clicable (id cells) but need to do back end for action on click.