Fixed login/out after Android app Authorizatoin meesup and other adjustments
Applied new API key to all scripts to get rid of new errors after adding android app to firebase,

updated firebase.js script source in headers to the new 4.6.2 version,

Changed account page name to Add Patient accross all pages - this simplifies things in that we dont need to make another page for that functionality and all that was included in the "account:" page was the details of the person currently logged in.  getting rid of current user info or moving it to another page saves time. will decide what to actually do with user info on next meeting.

Commented out javascript source in about and contact pages as there is nothing going on there and the script was causing an error to pop up in console, the error is controllable as in known harmless issue but just to simplifie it, it decided to comment js source tag out.